moves images array from ImageGallery.js into images.js
