+++ src/ImageGallery.js
@@ -1,33 +1,6 @@
 import React, { Component } from 'react';
 
-// import images from './images';
-
-const images = [
-  {
-    id: 1,
-    title: 'Cute Bunny',
-    description: 'Isn\'t it fuzzy-wuzzy cutest thing you\'ve ever seen?',
-    url: 'http://f.cl.ly/items/3g3J1G0w122M360w380O/3726490195_f7cc75d377_o.jpg'   
-  },
-  {
-    id: 2,
-    title: 'Snow Flake',
-    description: 'White Rabbit',
-    url: 'http://im5.leaderhero.com/wallpaper/20140328/12835c9a-f.jpg'
-  },
-  {
-    id: 3,
-    title: 'Floppy Bunny',
-    description: 'Can this floppy ear bunny fly?',
-    url: 'http://cdn.litlepups.net/2016/04/10/floppy-eared-bunny-365daysofdawn.jpg'
-  },
-  {
-    id: 4,
-    title: 'Crash Test Bunny',
-    description: 'This bunny is going bye bye for Delete test',
-    url: 'http://cdn.litlepups.net/2016/04/10/floppy-eared-bunny-365daysofdawn.jpg'
-  }, 
-];
+import images from './images';
 
 import DetailList from './DetailList';
 import GalleryList from './GalleryList';
